+++ core/src/inspect/column.rs
@@ -5,10 +5,34 @@ pub struct Column {
     pub is_primary: bool,
     pub name: String,
     pub oid: crate::pq::Oid,
+    #[deprecated(since = "3.3.0", note = "Use `Column::ty()` instead")]
     pub ty: String,
+    pub len: Option<i32>,
     pub default: Option<String>,
     pub is_notnull: bool,
     pub comment: Option<String>,
+}
+
+impl Column {
+    #[must_use]
+    pub fn ty(&self) -> String {
+        self.ty_recur(self.oid, self.len)
+    }
+
+    fn ty_recur(&self, oid: crate::pq::Oid, len: Option<i32>) -> String {
+        if let Ok(ty) = crate::pq::types::Type::try_from(oid) {
+            if let crate::pq::types::Kind::Array(oid) = ty.kind {
+                format!("{}[]", self.ty_recur(oid, len))
+            } else if let Some(len) = len {
+                format!("{}({len})", ty.name)
+            } else {
+                ty.name.to_string()
+            }
+        } else {
+            #[allow(deprecated)]
+            self.ty.clone()
+        }
+    }
 }
 
 /**
@@ -45,6 +69,7 @@ select
       when name.nspname != 'pg_catalog' then format('%s.%s', name.nspname, typ.typname)
       else typ.typname
     end as "ty",
+    nullif(att.atttypmod, -1) as "len",
     pg_catalog.pg_get_expr(def.adbin, def.adrelid) as "default",
     att.attnotnull as "is_notnull",
     dsc.description as "comment"