+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Adds `ltree`, `lquery` and `ltxtquery` types;
+- Impl `Copy` for `inspect::Type`;
 
 ## [3.2.0] - 2023-03-17