+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds `ltree`, `lquery` and `ltxtquery` types;
 - Impl `Copy` for `inspect::Type`;
 
+### Fixed
+- Bit type inspection;
+
 ## [3.2.0] - 2023-03-17
 
 ### Added